Fix bug where programming would try to unlock/erase too many sectors
There was an off by one error which I noticed when trying
to write an image that used a full 32k flash.

From before my change:
swddude -flash fargle.bin  -debug 2
 Notice..: Target supports 4 hardware breakpoints.
 Debug...:   Read program of 13404 bytes
 Debug...:   Unprotecting Flash sectors 0-4...

If you do the math, the sectors actually should be 0-3 for this
number of bytes.  I noticed the problem when the tool would die
while trying to unlock past the end of flash.